+++ cuvis_ai/node/__init__.py
@@ -61,7 +61,7 @@ from cuvis_ai.node.preprocessors import (
 )
 from cuvis_ai.node.spectral_angle_mapper import SpectralAngleMapper
 from cuvis_ai.node.spectral_extractor import BBoxSpectralExtractor
-from cuvis_ai.node.static_node import MaskPrompt
+from cuvis_ai.node.static_node import BBoxPrompt, MaskPrompt
 from cuvis_ai.node.video import (
     ToVideoNode,
     VideoFrameDataModule,
@@ -76,6 +76,7 @@ __all__ = [
     "BBoxesOverlayNode",
     "BBoxRoiCropNode",
     "BBoxSpectralExtractor",
+    "BBoxPrompt",
     "BinaryAnomalyLabelMapper",
     "ChannelNormalizeNode",
     "CameraEmulationFalseRGBSelector",